Feat: ChatServiceImpl 구현
- 사용자 참여 채팅방 목록 조회, 읽지 않은 채팅방만 필터링 조회
- 메시지 저장 및 읽음 처리 트랜잭션 적용
- ChatRoom -> ChatRoomListDto 변환 헬퍼 메서드 추가
- 존재하지 않는 채팅방/사용자 조회 시 예외 처리

+++ src/main/java/com/owl/trade_market/repository/ChatRepository.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 public interface ChatRepository extends JpaRepository<Chat, Long> {
 
     // 읽지 않은 메시지 수 카운트
-    long countByChatRoomIdAndUserIdNotAndIsReadIsFalse(Long chatRoomId, String currentUserId);
+    int countByChatRoomIdAndUserIdNotAndIsReadIsFalse(Long chatRoomId, String currentUserId);
 
     //상대방이 보낸 모든 메시지를 읽음으로 처리
     @Modifying
